+++ test/integration/verify/controls/alicloud_apsaradb_rds_instances.rb
@@ -1,11 +1,9 @@
-# frozen_string_literal: true
+alicloud_rds_db_id = input(:alicloud_rds_db_id, value: '', description: 'The Alicloud RDS DB identifier.')
 
-alicloud_rds_db_id = attribute(:alicloud_rds_db_id, value: '', description: 'The Alicloud RDS DB identifier.')
+title 'Test multiple AliCloud ApsaraDB RDS Instances'
 
-title 'Test multiple Alicloud ApsaraDB RDS Instances'
 control 'alicloud-apsaradb-rds-instances-1.0' do
-  impact 1.0
-  title 'Ensure Alicloud ApsaraDB RDS Instances have the correct properties.'
+  title 'Ensure AliCloud ApsaraDB RDS Instances have the correct properties.'
 
   describe alicloud_apsaradb_rds_instances do
     it { should exist }