+++ test/integration/verify/controls/alicloud_ecs_instance.rb
@@ -1,13 +1,10 @@
-# frozen_string_literal: true
-
 alicloud_instance_id = input(:alicloud_instance_id, value: '', description: 'AliCloud test instance ID.')
 alicloud_ram_role_name = input(:alicloud_ram_role_name, value: '', description: 'AliCloud RAM role name.')
 
 title 'Test single AliCloud ECS Instance'
 
 control 'alicloud-instance-1.0' do
-  impact 1.0
-  title 'Ensure Alicloud ECS Instance Class has correct attributes'
+  title 'Ensure AliCloud ECS Instance Class has correct attributes'
 
   describe alicloud_ecs_instance(instance_id: alicloud_instance_id) do # gets region from env var
     it { should exist }