+++ test/integration/verify/controls/alicloud_sts_caller_identity.rb
@@ -1,9 +1,6 @@
-# frozen_string_literal: true
-
 title 'Ensure AliCloud credentials being used for the Inspec scan have the correct properties.'
 
 control 'alicloud-sts-caller-identity-1.0' do
-  impact 1.0
   title 'Ensure AliCloud STS caller identity has the correct properties.'
 
   describe alicloud_sts_caller_identity do