+++ CHANGELOG.md
@@ -1,11 +1,26 @@
 # CHANGELOG
 
-<!-- latest_release 0.10.12 -->
+<!-- latest_release 0.10.15 -->
+## [v0.10.15](https://github.com/inspec/inspec-alicloud/tree/v0.10.15) (2022-12-13)
+
+#### Merged Pull Requests
+- RESOURCE-560 Remove the impact from the controls in inspec-alicloud [#96](https://github.com/inspec/inspec-alicloud/pull/96) ([soumyo13](https://github.com/soumyo13))
+<!-- latest_release -->
+
+## [v0.10.14](https://github.com/inspec/inspec-alicloud/tree/v0.10.14) (2022-12-06)
+
+#### Merged Pull Requests
+- RESOURCE-542 Remove frozen string literal from inspec alicloud [#95](https://github.com/inspec/inspec-alicloud/pull/95) ([soumyo13](https://github.com/soumyo13))
+
+## [v0.10.13](https://github.com/inspec/inspec-alicloud/tree/v0.10.13) (2022-12-02)
+
+#### Merged Pull Requests
+- RESOURCE-518 Add the missing Unit tests in inspec-alicloud [#94](https://github.com/inspec/inspec-alicloud/pull/94) ([soumyo13](https://github.com/soumyo13))
+
 ## [v0.10.12](https://github.com/inspec/inspec-alicloud/tree/v0.10.12) (2022-11-22)
 
 #### Merged Pull Requests
 - RESOURCE-521 Fixing the lint errors [#92](https://github.com/inspec/inspec-alicloud/pull/92) ([soumyo13](https://github.com/soumyo13))
-<!-- latest_release -->
 
 ## [v0.10.11](https://github.com/inspec/inspec-alicloud/tree/v0.10.11) (2022-11-11)